Move docker folder to root
