+++ docker/zeppelin-site.xml
@@ -1,0 +1,12 @@
+<configuration>
+    <property>
+        <name>zeppelin.server.jetty.request.header.size</name>
+        <value>16000</value>
+        <description>Http Request Header Size Limit (to prevent HTTP 413)</description>
+    </property>
+    <property>
+        <name>zeppelin.anonymous.allowed</name>
+        <value>false</value>
+        <description>Anonymous user allowed by default</description>
+    </property>
+</configuration>
