--- zeppelin/docker/shiro.ini
@@ -1,34 +1,0 @@
-[users]
-admin = admin, admin
-user1 = user1, role1, role2
-user2 = user2, role3
-user3 = user3, role2
-
-[main]
-sessionManager = org.apache.shiro.web.session.mgt.DefaultWebSessionManager
-
-cookie = org.apache.shiro.web.servlet.SimpleCookie
-cookie.name = JSESSIONID
-cookie.httpOnly = true
-
-sessionManager.sessionIdCookie = $cookie
-securityManager.sessionManager = $sessionManager
-
-# 86,400,000 milliseconds = 24 hour
-securityManager.sessionManager.globalSessionTimeout = 86400000
-shiro.loginUrl = /api/login
-
-[roles]
-role1 = *
-role2 = *
-role3 = *
-admin = *
-
-[urls]
-#/api/version = anon
-/api/interpreter/setting/restart/** = authc
-/api/interpreter/** = authc, roles[admin]
-/api/configurations/** = authc, roles[admin]
-/api/credential/** = authc, roles[admin]
-#/** = anon
-/** = authc
